+++ tests/RSetTest.php
@@ -428,6 +428,52 @@ class RSetTest extends PHPUnit_Framework_TestCase
 	}
 
 ///////////////////////////////////////////////////////////////////////////////
+// GetOccurrences
+
+	public function testGetOccurrences()
+	{
+		$rset = new RSet();
+		$rset->addRRule(new RRule(array(
+			'FREQ' => 'DAILY',
+			'DTSTART' => '2017-01-01'
+		)));
+
+		$this->assertCount(1, $rset->getOccurrences(1));
+		$this->assertEquals([date_create('2017-01-01')], $rset->getOccurrences(1));
+		$this->assertCount(5, $rset->getOccurrences(5));
+		$this->assertEquals([
+			date_create('2017-01-01'),date_create('2017-01-02'),date_create('2017-01-03'),
+			date_create('2017-01-04'),date_create('2017-01-05')
+		], $rset->getOccurrences(5));
+		try {
+			$rset->getOccurrences();
+			$this->fail('Expected exception (infinite rule) not thrown');
+		} catch ( \LogicException $e ) {
+
+		}
+	}
+
+	public function testGetOccurrencesBetween()
+	{
+		$rset = new RSet();
+		$rset->addRRule(new RRule(array(
+			'FREQ' => 'DAILY',
+			'DTSTART' => '2017-01-01'
+		)));
+
+		$this->assertCount(1, $rset->getOccurrencesBetween('2017-01-01', null, 1));
+		$this->assertCount(1, $rset->getOccurrencesBetween('2017-02-01', '2017-12-31', 1));
+		$this->assertEquals([date_create('2017-02-01')], $rset->getOccurrencesBetween('2017-02-01', '2017-12-31', 1));
+		$this->assertCount(5, $rset->getOccurrencesBetween('2017-01-01', null, 5));
+		try {
+			$rset->getOccurrencesBetween('2017-01-01', null);
+			$this->fail('Expected exception (infinite rule) not thrown');
+		} catch ( \LogicException $e ) {
+			
+		}
+	}
+
+///////////////////////////////////////////////////////////////////////////////
 // RFC Strings
 
 	public function rfcStrings()